requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-chroma
pypdf
sentence-transformers
google-generativeai
python-dotenv
fastapi
uvicorn
requests
python-multipart